+++ mayan/apps/document_states/handlers.py
@@ -37,18 +37,26 @@ def handler_trigger_transition(sender, **kwargs):
         app_label='document_states', model_name='WorkflowTransition'
     )
 
-    trigger_transitions = WorkflowTransition.objects.filter(trigger_events__event_type__name=kwargs['instance'].verb)
+    trigger_transitions = WorkflowTransition.objects.filter(
+        trigger_events__event_type__name=kwargs['instance'].verb
+    )
 
     if isinstance(action.target, Document):
-        workflow_instances = WorkflowInstance.objects.filter(workflow__transitions__in=trigger_transitions, document=action.target).distinct()
+        workflow_instances = WorkflowInstance.objects.filter(
+            workflow__transitions__in=trigger_transitions, document=action.target
+        ).distinct()
     elif isinstance(action.action_object, Document):
-        workflow_instances = WorkflowInstance.objects.filter(workflow__transitions__in=trigger_transitions, document=action.action_object).distinct()
+        workflow_instances = WorkflowInstance.objects.filter(
+            workflow__transitions__in=trigger_transitions, document=action.action_object
+        ).distinct()
     else:
         workflow_instances = WorkflowInstance.objects.none()
 
     for workflow_instance in workflow_instances:
         # Select the first transition that is valid for this workflow state
-        valid_transitions = list(set(trigger_transitions) & set(workflow_instance.get_transition_choices()))
+        valid_transitions = list(
+            set(trigger_transitions) & set(workflow_instance.get_transition_choices())
+        )
         if valid_transitions:
             workflow_instance.do_transition(
                 comment=_('Event trigger: %s') % EventType.get(name=action.verb).label,