+++ HISTORY.rst
@@ -296,6 +296,8 @@
 * Add Latvian translation.
 * Support search model selection.
 * Support passing a queryset factory to the search model.
+* Add workflow actions to grant or remove permissions to
+  a document.
 
 3.1.11 (2019-04-XX)
 ===================